Rather than a long we will now use a  -/+ 2^53 double to have the same type as the Mongo shell. We will not use boxing.

+++ orfiles/src/main/java/org/objectrepository/io/OrFilesFactory.java
@@ -170,8 +170,8 @@ public abstract class OrFilesFactory implements OrFiles {
         this.shardKey = shardKey;
     }
 
-    public long getS() {
-        return (shardKey == null || shardKey.isEmpty()) ? (long) new Random().nextInt() : Long.parseLong(shardKey);
+    public double getS() {
+        return (shardKey == null || shardKey.isEmpty()) ?  new Random().nextInt() : Double.parseDouble(shardKey);
     }
 
     public String getPid() {
